[SCRUM-37] As a user I can see discounts

+++ src/main/java/com/sushi/wasabi/controller/FoodItemController.java
@@ -22,7 +22,7 @@ public Page<FoodItem> getAllFoods(
 }
 
 @GetMapping("/{id}")
-public FoodItem getFoodById(@PathVariable Long id){
+public FoodItem getFoodById(@PathVariable Integer id){
     return foodItemService.getById(id);
 }
 }
